add Ordering.io module and services
