+++ .env
@@ -8,6 +8,6 @@ DATABASE_URL="postgresql://postgres:huy123@localhost:5432/gomunchi?schema=public
 API_KEY="XCc6ccCL3ziT4kLV6CmAY0k1kODjLK6aVjXayv0bovIZoV7Rh6D0_SFxaA_vTEOel"
 PROJECT_NAME="peperoni"
 BASE_URL="https://apiv4.ordering.co/v400/en/${PROJECT_NAME}"
-CLIENT_URL_DEVELOPEMENT = "http://localhost:5000"
+CLIENT_URL_DEVELOPEMENT = "http://localhost:3000"
 CLIENT_URL_PRODUCTION = ""
 JWT_SECRET="secret"
