+++ package.json
@@ -33,6 +33,7 @@
     "@nestjs/passport": "^9.0.0",
     "@nestjs/platform-express": "^9.0.0",
     "@nestjs/platform-socket.io": "^9.2.1",
+    "@nestjs/swagger": "^6.1.4",
     "@nestjs/websockets": "^9.2.1",
     "@prisma/client": "^4.6.1",
     "@socket.io/redis-adapter": "^7.2.0",
