+++ src/auth/auth.controller.ts
@@ -1,19 +1,20 @@
 import { Controller, Post, Body } from '@nestjs/common';
+import { OrderingIoService } from 'src/ordering.io/ordering.io.service';
 import { AuthCredentials } from 'src/type';
-import { AuthService } from './auth.service';
+
 
 @Controller('auth')
 export class AuthController {
-  constructor(private authService: AuthService) {}
+  constructor( private orderingIo: OrderingIoService) {}
 
   @Post('signup')
   signup(@Body() credentials: AuthCredentials) {
     console.log(credentials);
-    return this.authService.signup(credentials);
+    return this.orderingIo.signUp(credentials);
   }
 
   @Post('signin')
   signin(@Body() credentials: AuthCredentials) {
-    return this.authService.signin(credentials);
+    return this.orderingIo.signIn(credentials);
   }
 }
