+++ src/auth/auth.module.ts
@@ -1,13 +1,14 @@
 import { Module } from '@nestjs/common';
 import { JwtModule } from '@nestjs/jwt';
+import { OrderingIoModule } from 'src/ordering.io/ordering.io.module';
+import { OrderingIoService } from 'src/ordering.io/ordering.io.service';
 import { UtilsModule } from 'src/utils/utils.module';
 import { AuthController } from './auth.controller';
-import { AuthService } from './auth.service';
+
 import { JwtStrategy } from './strategy/jwt.strategy';
 
 @Module({
-  imports: [JwtModule.register({}), UtilsModule],
+  imports: [ UtilsModule, OrderingIoModule],
   controllers: [AuthController],
-  providers: [AuthService, JwtStrategy],
 })
 export class AuthModule {}
