--- src/business/business.service..ts
@@ -1,95 +1,0 @@
-import { ForbiddenException, Injectable } from '@nestjs/common';
-import { plainToClass } from 'class-transformer';
-import { AllBusinessDto, BusinessDto } from './dto/business.dto';
-import axios from 'axios';
-import { UtilsService } from 'src/utils/utils.service';
-@Injectable()
-export class BusinessService {
-  constructor(private utils: UtilsService) {}
-  async getAllBusiness(accessToken: string) {
-    const options = {
-      method: 'GET',
-      url: `${this.utils.getEnvUrl('business')}?type=1&params=zones%2Cname&mode=dashboard`,
-      headers: {
-        accept: 'application/json',
-        Authorization: `Bearer ${accessToken}`,
-      },
-    };
-    try {
-      const response = await axios.request(options);
-      const businessResponseObject = response.data.result;
-      const businessResponse = plainToClass(AllBusinessDto, businessResponseObject);
-      console.log(businessResponse);
-      return businessResponse;
-    } catch (error) {
-      const errorMsg = error.response.data.result;
-      throw new ForbiddenException(errorMsg);
-    }
-  }
-  async getBusinessById(businessId: number, accessToken: string) {
-    console.log(businessId);
-    const options = {
-      method: 'GET',
-      url: `${this.utils.getEnvUrl('business', businessId)}?mode=dashboard`,
-      headers: {
-        accept: 'application/json',
-        Authorization: `Bearer ${accessToken}`,
-      },
-    };
-    console.log(`${this.utils.getEnvUrl('business', businessId)}?mode=dashboard`);
-    try {
-      const response = await axios.request(options);
-      const businessResponseObject = response.data.result;
-      const businessResponse = plainToClass(BusinessDto, businessResponseObject);
-      console.log(businessResponse);
-      return businessResponse;
-    } catch (error) {
-      const errorMsg = error.response.data.result;
-      throw new ForbiddenException(errorMsg);
-    }
-  }
-
-  async getBusinessOnline(businessId: number, accessToken: string) {
-    const options = {
-      method: 'POST',
-      url: `${this.utils.getEnvUrl('business', businessId)}`,
-      data: { enabled: true },
-      headers: {
-        accept: 'application/json',
-        Authorization: `Bearer ${accessToken}`,
-      },
-    };
-    try {
-      const response = await axios.request(options);
-      const businessResponseObject = response.data.result;
-
-      console.log(businessResponseObject);
-      return `Business Online`;
-    } catch (error) {
-      const errorMsg = error.response.data.result;
-      throw Error(errorMsg);
-    }
-  }
-
-  async getBusinessOffline(businessId: number, accessToken: string) {
-    const options = {
-      method: 'POST',
-      url: `${this.utils.getEnvUrl('business', businessId)}`,
-      data: { enabled: false },
-      headers: {
-        accept: 'application/json',
-        Authorization: `Bearer ${accessToken}`,
-      },
-    };
-    try {
-      const response = await axios.request(options);
-      const businessResponseObject = response.data.result;
-
-      console.log(businessResponseObject);
-      return `Business Offline`;
-    } catch (error) {
-      const errorMsg = error.response.data.result;
-      throw Error(errorMsg);
-    }
-  }
-}
