+++ src/main.ts
@@ -1,17 +1,36 @@
 import { ValidationPipe } from '@nestjs/common';
 import { NestFactory } from '@nestjs/core';
-import { toASCII } from 'punycode';
+import { DocumentBuilder } from '@nestjs/swagger';
+import { SwaggerDocumentOptions, SwaggerModule } from '@nestjs/swagger/dist';
+
 import { AppModule } from './app/app.module';
 declare const module: any;
 
 async function bootstrap() {
   const app = await NestFactory.create(AppModule);
+  const config = new DocumentBuilder()
+    .setTitle('Api documentation')
+    .setDescription('The API description of munchi-apis')
+    .setVersion('1.0')
+    .addTag('Munchi-apis')
+    .build();
+    const options: SwaggerDocumentOptions =  {
+      deepScanRoutes: true,
+      operationIdFactory: (
+        controllerKey: string,
+        methodKey: string
+      ) => methodKey
+    };
+  const document = SwaggerModule.createDocument(app, config,options);
+  SwaggerModule.setup('api', app, document);
   app.enableCors({
-    origin: [`${process.env.CLIENT_URL_DEVELOPEMENT}`]
-  })
-  app.useGlobalPipes(new ValidationPipe({
-    whitelist:true
-  }));
+    origin: [`${process.env.CLIENT_URL_DEVELOPEMENT}`],
+  });
+  app.useGlobalPipes(
+    new ValidationPipe({
+      whitelist: true,
+    }),
+  );
   await app.listen(5000);
   if (module.hot) {
     module.hot.accept();
