+++ src/user/user.controller.ts
@@ -1,16 +1,16 @@
 import { Controller, Get, Param, Request, UseGuards } from '@nestjs/common';
 import { JwtGuard } from 'src/auth/guard';
+import { OrderingIoService } from 'src/ordering.io/ordering.io.service';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { OrderId } from 'src/type';
-import { UserService } from './user.service';
 
 @Controller('user')
 export class UserController {
-  constructor(private userService: UserService, private prisma: PrismaService) {}
+  constructor(private orderingIo: OrderingIoService, private prisma: PrismaService) {}
   @UseGuards(JwtGuard)
   @Get(':userId')
   async getUser(@Param('userId') userId: OrderId, @Request() req) {
     const { id } = req.user;
-    return this.userService.getUser(id);
+    return this.orderingIo.getUser(id);
   }
 }
